+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All important changes to `fiisoft-logger` will be documented in this file
 
+## 1.1.0
+
+ - added InstantSmartLogger and InstantLogsMonitor.
+ - added AbstractLogsMonitor
+ - DefaultLogsMonitor inherits from AbstractLogsMonitor 
+ - DefaultSmartLogger is not final
+
 ## 1.0.0
 
 Initial version of library.